文章和问答模块接口已实现，待前端联调

+++ service/service_order/src/main/java/com/atguigu/orderservice/service/impl/TPayLogServiceImpl.java
@@ -91,7 +91,7 @@ public class TPayLogServiceImpl extends ServiceImpl<TPayLogMapper, TPayLog> impl
         QueryWrapper<TOrder> wrapper = new QueryWrapper<>();
         wrapper.eq("order_no",orderNo);
         TOrder order = orderService.getOne(wrapper);
-        if(order.getStatus().intValue() == 1) return;
+        if(order.getStatus() == 1) return;
         order.setStatus(1);
         orderService.updateById(order);
